Fix some FindBugs bugs (array hashcode, unclosed connection/stream and int division for float comparison).
git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@23256 219f1c6e-4a0e-43c8-96f2-6e749be13fd4

+++ modules/kerberos/src/main/java/com/l7tech/kerberos/KerberosTicketRepository.java
@@ -234,7 +234,7 @@ public class KerberosTicketRepository {
     /**
      * Imposed limit on the cache such that it doesn't fill up the ssg memory
      */
-    private static final int CACHE_SIZE_LIMIT = 250;
+    private static final float CACHE_SIZE_LIMIT = 250;
 
     /**
      * Threshold to be reached that kicks off the cleanup task
